requirements.txt
# requirements.txt

# Core versions
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
ipykernel==6.29.3
matplotlib==3.8.3
zarr==2.17.0
hydra-core==1.3.2
einops==0.6.1
tqdm==4.66.4
h5py==3.7.0
gymnasium==0.29.1
pymunk==6.7.0
moviepy==1.0.3
imageio==2.17.0
imageio-ffmpeg==0.4.7
diffusers==0.34.0

# From Git or custom sources
git+https://github.com/facebookresearch/r3m.git
aeon==1.0.0
roma==1.5.2.1
open3d
wandb
pygame
robosuite @ https://github.com/ARISE-Initiative/robosuite/archive/b9d8d3de5e3dfd1724f4a0e6555246c460407daa.tar.gz
robomimic @ https://github.com/pointW/robomimic/archive/8aad5b3caaaac9289b1504438a7f5d3a76d06c07.tar.gz
